Tweaks from code review: remove TaskDetails, make getInstance includingCreating clearer, expanded imports
